+++ yash-cli/CHANGELOG.md
@@ -9,6 +9,16 @@ used by other programs.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.3.0] - Unreleased
+
+### Added
+
+- The `portable` shell option, which restricts the shell to portable constructs
+  so that scripts can be checked for portability. See
+  [Writing portable scripts](https://magicant.github.io/yash-rs/posix.html#writing-portable-scripts)
+  in the manual for the constructs it rejects. More checks will be added in
+  future releases.
+
 ## [3.2.1] - 2026-06-21
 
 ### Changed
@@ -361,6 +371,7 @@ later.
 
 - Initial release of the shell
 
+[3.3.0]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-3.3.0
 [3.2.1]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-3.2.1
 [3.2.0]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-3.2.0
 [3.1.0]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-3.1.0